+++ src/Report/StockSummary.php
@@ -205,4 +205,3 @@ class StockSummary extends AbstractReport
 		}
 	}
 }
-